settings: link the exchange disable flag
+ apply settings to all exchanges

but remove the btc withdraw address from valid keys by default, security issue...

+++ web/yaamp/core/trading/empoex_trading.php
@@ -5,6 +5,8 @@ function doEmpoexTrading($quick=false)
 	$exchange = 'empoex';
 	$updatebalances = true;
 
+	if (exchange_get($exchange, 'disabled')) return;
+
 	$flushall = rand(0, 8) == 0;
 	if($quick) $flushall = false;
